Type Alias sc_subspace_chain_specs::ExecutionChainSpec

pub type ExecutionChainSpec<ExecutionGenesisConfig, Ext = NoExtension> = SerializableChainSpec<ExecutionGenesisConfig, Ext>;

Specialized ChainSpec for the execution runtime.

Aliased Type

struct ExecutionChainSpec<ExecutionGenesisConfig, Ext = NoExtension> { /* private fields */ }

Implementations

impl<GenesisConfig, Extensions> SerializableChainSpec<GenesisConfig, Extensions>where GenesisConfig: RuntimeGenesis + 'static, Extensions: GetExtension + Serialize + Clone + Send + Sync + 'static,

pub fn boot_nodes(&self) -> &[MultiaddrWithPeerId]

A list of bootnode addresses.

pub fn name(&self) -> &str

Spec name.

pub fn id(&self) -> &str

Spec id.

pub fn telemetry_endpoints(&self) -> &Option<TelemetryEndpoints>

Telemetry endpoints (if any)

pub fn protocol_id(&self) -> Option<&str>

Network protocol id.

pub fn fork_id(&self) -> Option<&str>

Optional network fork identifier.

pub fn properties(&self) -> Properties

Additional loosly-typed properties of the chain.

Returns an empty JSON object if ‘properties’ not defined in config

pub fn add_boot_node(&mut self, addr: MultiaddrWithPeerId)

Add a bootnode to the list.

pub fn extensions(&self) -> &Extensions

Returns a reference to the defined chain spec extensions.

pub fn extensions_mut(&mut self) -> &mut Extensions

Returns a mutable reference to the defined chain spec extensions.

pub fn from_genesis<F: Fn() -> GenesisConfig + 'static + Send + Sync>( name: &str, id: &str, chain_type: ChainType, constructor: F, boot_nodes: Vec<MultiaddrWithPeerId>, telemetry_endpoints: Option<TelemetryEndpoints>, protocol_id: Option<&str>, fork_id: Option<&str>, properties: Option<Properties>, extensions: Extensions ) -> Self

Create hardcoded spec.

impl<GenesisConfig, Extensions> SerializableChainSpec<GenesisConfig, Extensions>where Extensions: DeserializeOwned,

pub fn from_json_bytes( json: impl Into<Cow<'static, [u8]>> ) -> Result<Self, String>

Parse json content into a ChainSpec

pub fn from_json_file(path: PathBuf) -> Result<Self, String>

Parse json file into a ChainSpec